Update 76_style_guide.md
I was told that each page without a first-level header does not get compiled. It might be the same for pages without status in the header line. 
Please only accept the change if this is true, and tell me if this is not true (I'll exclude such pages from knowledge graph creation as well).